api-ai/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
langchain-groq==0.1.4
langgraph==0.0.41
google-api-python-client==2.126.0
google-auth==2.29.0
python-dotenv==1.0.1
transformers==4.41.1
